# docs/swagger.yaml
basePath: /api/v1
definitions:
  docs.ErrorResponse:
    properties:
      details: {}
      error:
        example: ERROR_CODE
        type: string
      message:
        example: An error occurred
        type: string
      success:
        example: false
        type: boolean
    type: object
  docs.SuccessResponse:
    properties:
      data: {}
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  go-futsal-booking-api_internal_dto_request.CreateScheduleRequest:
    properties:
      day_of_week:
        type: integer
      end_time:
        type: string
      field_id:
        type: integer
      price:
        type: number
      start_time:
        type: string
    required:
    - day_of_week
    - end_time
    - field_id
    - price
    type: object
...
  go-futsal-booking-api_internal_dto_request.UpdateScheduleRequest:
    properties:
      day_of_week:
        type: integer
      end_time:
        type: string
      field_id:
        type: integer
      price:
        type: number
      start_time:
        type: string
    required:
    - day_of_week
    - end_time
    - field_id
    - price
    type: object
...
  go-futsal-booking-api_internal_dto_response.ScheduleResponse:
    properties:
      created_at:
        type: string
      day_of_week:
        type: integer
      end_time:
        type: string
      field_id:
        type: integer
      id:
        type: integer
      price:
        type: number
      start_time:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: This is a REST API for a Futsal Booking application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Futsal Booking API
  version: "1.0"
paths:
...
  /schedules:
    get:
      description: Get a list of all schedules associated with a specific field
      parameters:
      - description: Field ID
        in: query
        name: fieldId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Schedule retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/docs.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/go-futsal-booking-api_internal_dto_response.ScheduleResponse'
                  type: array
              type: object
        "400":
          description: Missing or Invalid Field ID
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized (Missing Token)
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Field Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get schedules by field ID
      tags:
      - Schedules
    post:
      consumes:
      - application/json
      description: Create a new weekly schedule for a specific field
      parameters:
      - description: Schedule creation request
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/go-futsal-booking-api_internal_dto_request.CreateScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Schedule successfully created
          schema:
            allOf:
            - $ref: '#/definitions/docs.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/go-futsal-booking-api_internal_dto_response.ScheduleResponse'
              type: object
        "400":
          description: Bad Request or Validation Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized (Missing Token)
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "403":
          description: Forbidden (Not Admin)
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Field Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new schedule (Admin only)
      tags:
      - Schedules
  /schedules/{id}:
    delete:
      description: Delete an existing schedule by its ID
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Schedule deleted successfully
          schema:
            $ref: '#/definitions/docs.SuccessResponse'
        "400":
          description: Invalid Schedule ID
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized (Missing Token)
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "403":
          description: Forbidden (Not Admin)
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Schedule Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a schedule (Admin only)
      tags:
      - Schedules
    get:
      description: Get details of a specific schedule by its ID
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Schedule retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/docs.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/go-futsal-booking-api_internal_dto_response.ScheduleResponse'
              type: object
        "400":
          description: Invalid Schedule ID
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized (Missing Token)
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Schedule Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get schedule by ID
      tags:
      - Schedules
    put:
      consumes:
      - application/json
      description: Update details of an existing schedule
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: integer
      - description: Schedule update request
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/go-futsal-booking-api_internal_dto_request.UpdateScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Schedule successfully updated
          schema:
            allOf:
            - $ref: '#/definitions/docs.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/go-futsal-booking-api_internal_dto_response.ScheduleResponse'
              type: object
        "400":
          description: Bad Request, Invalid ID, or Validation Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized (Missing Token)
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "403":
          description: Forbidden (Not Admin)
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Schedule Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a schedule (Admin only)
      tags:
      - Schedules
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"